fix(changelog): improve word wrap and line length config/flags
Currently lines are wrapped based on the description only but not on the scope.
Now the word-wrap is part of the template and takes the full line into account.

```hbs
{{#word-wrap}}
..
{{/word-wrap}}
```

This also add control for word wrap and line length as flags.

`--no-wrap` to disable wrapping.
`--line-length <LINE_LENGTH>` to set the line length.

Refs: #84