feat(commit): save message to a file to recover or use as a git editor
This save the convco commit message to a file `.git/CONVCO_MSG`.
In case a pre-commit hook failed, the next run of `convco commit` would recover from this file.
`convco commit` can now also be used as a git editor.
In this case git will call convco commit and not the other way around.

Refs: #157